pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-wiki"
version = "0.1.0"
description = "Karpathy's LLM Wiki pattern implemented on Databricks"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "databricks-sdk>=0.40.0",
    "databricks-sql-connector>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg[binary,pool]>=3.2.0",
    "loguru>=0.7.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "markupsafe>=2.1.0",
    "markdown>=3.6",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_wiki"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
